pyproject.toml
 [project]
 name = "api-comerce-in"
 version = "0.1.0"
 description = "E-Commerce API (FastAPI) backend"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
  "alembic>=1.16",
   "bcrypt>=4.0",
   "email-validator>=2.0",
   "fastapi>=0.115",
   "pydantic-settings>=2.6",
   "python-dotenv>=1.0",
   "python-multipart>=0.0.9",
   "python-jose[cryptography]>=3.3",
  "psycopg[binary]>=3.2",
  "sqlalchemy>=2.0",
   "uvicorn[standard]>=0.30",
 ]
 
 [project.optional-dependencies]
 dev = [
   "httpx>=0.27",
   "pytest>=8.0",
   "ruff>=0.6",
 ]
 
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
 
